To do left 3 things from paper.

+++ src/main/scala/hr/fer/ztel/dipl/model/ItemCustomer.scala
@@ -1,5 +1,5 @@
 package hr.fer.ztel.dipl.model
 
-case class ItemCustomerRecord(itemId : Int, customerId : Int, quantity : Float)
+case class ItemCustomerRecord(itemId : Int, customerId : Int, quantity : Double)
 
 case class ItemCustomer(itemId : Int, customerId : Int)
